requirements.txt
requests==2.32.3
python-dotenv==1.1.0
waitress==3.0.2
Flask==3.1.1
gunicorn==23.0.0